Removing duplicate code from MessageTransaction
